added arrow metadata + refactored code
