+++ crates/viewer/re_dataframe_ui/src/lib.rs
@@ -3,6 +3,7 @@
 mod datafusion_adapter;
 mod datafusion_table_widget;
 mod display_record_batch;
+mod header_tooltip;
 mod requested_object;
 mod table_blueprint;
 pub mod table_utils;
